backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
python-multipart
websockets
httpx
google-auth
PyJWT
